Move to right namespace, change constructor
