+++ tests/Routes/Wiki/CreateTest.php
@@ -7,7 +7,7 @@ use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use App\User;
 use Illuminate\Support\Facades\Queue;
-use App\Jobs\ElasticSearch\ElasticSearchIndexInit;
+use App\Jobs\CirrusSearch\ElasticSearchIndexInit;
 use App\Jobs\ProvisionWikiDbJob;
 use App\Jobs\MediawikiInit;
 use App\WikiSetting;
